[HOTFIX] Fix non-systematic calculation returning the same results
Fixes #9

After 021e7d212d8080cf623e292e3ed44dfdb0914493, if one would use non-systematic mode, then the result would be that the first place has 100% COBEN and everyone else has immunity. I failed to account that the first argument to `shuffleIter` can be an iterator and not an array, so I changed that function to account for that. I also updated the documentation.